+++ CHANGELOG.md
@@ -10,7 +10,34 @@
 - N/A
 
 ## 🐞 Fixes
-- N/A
+- Fixed hitboxes for the following entities being incorrectly sized:
+  - Aberrant Chestburster
+  - Aberrant Drone
+  - Aberrant Facehugger
+  - Aberrant Ovamorph
+  - Aberrant Praetorian
+  - Aberrant Queen
+  - Aberrant Warrior
+  - Irradiated Drone
+  - Irradiated Praetorian
+  - Irradiated Queen
+  - Irradiated Warrior
+  - Nether Chestburster
+  - Nether Drone
+  - Nether Facehugger
+  - Nether Ovamorph
+  - Nether Praetorian
+  - Nether Queen
+  - Nether Warrior
+  - Royal Aberrant Chestburster
+  - Royal Aberrant Facehugger
+  - Royal Aberrant Ovamorph
+  - Royal Nether Chestburster
+  - Royal Nether Facehugger
+  - Royal Nether Ovamorph
+  - Royal Chestburster
+  - Royal Facehugger
+  - Royal Ovamorph
 
 ## 🛠 Data Pack
 - N/A